project/requirements.txt
Flask
Flask-Session
requests
psycopg2-binary
flask-cors